pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["auramaur*", "config*"]

[project]
name = "auramaur"
version = "0.1.0"
description = "Polymarket prediction market trading bot powered by Claude NLP"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.1.8",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiohttp>=3.9",
    "aiosqlite<=0.17.0",
    "asyncpraw>=7.7",
    "tweepy>=4.14",
    "fredapi>=0.5",
    "newsapi-python>=0.2.7",
    "ddgs>=7.0",
    "feedparser>=6.0",
    "rich>=13.0",
    "structlog>=24.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
kalshi = ["kalshi-python>=2.1.0"]
ibkr = ["ib_async>=1.0.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
auramaur = "auramaur.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
